--- src/ReveationLabs.CartAbandonmentReminder.Core/ReveationLabs.CartAbandonmentReminder.Core.csproj
@@ -1,25 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-    <TargetFramework>net6.0</TargetFramework>
-  </PropertyGroup>
-
-  <PropertyGroup>
-      <!-- Project is not a test project -->
-      <SonarQubeTestProject>false</SonarQubeTestProject>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <Folder Include="Events\" />
-    <Folder Include="Models\" />
-    <Folder Include="Notifications\" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <PackageReference Include="Microsoft.SourceLink.GitHub" Version="1.0.0" PrivateAssets="All" />
-    <PackageReference Include="VirtoCommerce.CartModule.Core" Version="3.201.0" />
-    <PackageReference Include="VirtoCommerce.StoreModule.Core" Version="3.206.0" />
-    <PackageReference Include="VirtoCommerce.Platform.Core" Version="3.200.0" />
-  </ItemGroup>
-
-</Project>
