--- src/ReveationLabs.CartAbandonmentReminder.Core/Services/IExtendShoppingCartSearchService.cs
@@ -1,14 +1,0 @@
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-using VirtoCommerce.CartModule.Core.Model.Search;
-
-namespace ReveationLabs.CartAbandonmentReminder.Core.Services
-{
-    public interface IExtendShoppingCartSearchService
-    {
-        Task<ShoppingCartSearchResult> SearchCartAsync(ShoppingCartSearchCriteria criteria);
-    }
-}
