--- src/ReveationLabs.CartAbandonmentReminder.Data/Repositories/DesignTimeDbContextFactory.cs
@@ -1,17 +1,0 @@
-using Microsoft.EntityFrameworkCore;
-using Microsoft.EntityFrameworkCore.Design;
-
-namespace ReveationLabs.CartAbandonmentReminder.Data.Repositories
-{
-    public class DesignTimeDbContextFactory : IDesignTimeDbContextFactory<CartAbandonmentReminderDbContext>
-    {
-        public CartAbandonmentReminderDbContext CreateDbContext(string[] args)
-        {
-            var builder = new DbContextOptionsBuilder<CartAbandonmentReminderDbContext>();
-
-            builder.UseSqlServer("Data Source=(local);Initial Catalog=VirtoCommerce3;Persist Security Info=True;User ID=virto;Password=virto;MultipleActiveResultSets=True;Connect Timeout=30");
-
-            return new CartAbandonmentReminderDbContext(builder.Options);
-        }
-    }
-}
