--- src/ReveationLabs.CartAbandonmentReminder.Data/ReveationLabs.CartAbandonmentReminder.Data.csproj
@@ -1,37 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-    <TargetFramework>net6.0</TargetFramework>
-  </PropertyGroup>
-
-  <PropertyGroup>
-      <!-- Project is not a test project -->
-      <SonarQubeTestProject>false</SonarQubeTestProject>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <ProjectReference Include="..\ReveationLabs.CartAbandonmentReminder.Core\ReveationLabs.CartAbandonmentReminder.Core.csproj" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <Folder Include="Caching\" />
-    <Folder Include="ExportImport\" />
-    <Folder Include="Handlers\" />
-    <Folder Include="Models\" />
-    <Folder Include="Repositories\" />
-    <Folder Include="Services\" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <PackageReference Include="Hangfire" Version="1.7.14" />
-    <PackageReference Include="Microsoft.EntityFrameworkCore.Tools" Version="6.0.0">
-      <PrivateAssets>all</PrivateAssets>
-      <IncludeAssets>runtime; build; native; contentfiles; analyzers; buildtransitive</IncludeAssets>
-    </PackageReference>
-    <PackageReference Include="Microsoft.SourceLink.GitHub" Version="1.0.0" PrivateAssets="All" />
-    <PackageReference Include="VirtoCommerce.CartModule.Core" Version="3.201.0" />
-    <PackageReference Include="VirtoCommerce.CartModule.Data" Version="3.201.0" />
-    <PackageReference Include="VirtoCommerce.Platform.Data" Version="3.239.0" />
-  </ItemGroup>
-
-</Project>
