--- src/ReveationLabs.CartAbandonmentReminder.Web/ReveationLabs.CartAbandonmentReminder.Web.csproj
@@ -1,22 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk.Web">
-
-  <PropertyGroup>
-    <TargetFramework>net6.0</TargetFramework>
-    <OutputType>Library</OutputType>
-  </PropertyGroup>
-
-  <PropertyGroup>
-      <!-- Project is not a test project -->
-      <SonarQubeTestProject>false</SonarQubeTestProject>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <PackageReference Include="VirtoCommerce.StoreModule.Core" Version="3.206.0" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <ProjectReference Include="..\ReveationLabs.CartAbandonmentReminder.Core\ReveationLabs.CartAbandonmentReminder.Core.csproj" />
-    <ProjectReference Include="..\ReveationLabs.CartAbandonmentReminder.Data\ReveationLabs.CartAbandonmentReminder.Data.csproj" />
-  </ItemGroup>
-
-</Project>
