--- src/ReveationLabs.CartAbandonmentReminder.Web/package.json
@@ -1,18 +1,0 @@
-{
-    "version": "1.0.0",
-    "name": "asp.net",
-    "private": true,
-    "scripts": {
-        "webpack:dev": "webpack --mode=development",
-        "webpack:watch": "webpack --watch --mode=development",
-        "webpack:build": "webpack --mode=production"
-    },
-    "devDependencies": {
-        "clean-webpack-plugin": "^1.0.0",
-        "css-loader": "^1.0.1",
-        "mini-css-extract-plugin": "^0.4.5",
-        "webpack": "^5.58.1",
-        "webpack-cli": "^4.9.0"
-    },
-    "dependencies": {}
-}
