+++ src/Sharpcode.CartAbandonmentReminder.Web/Localizations/en.CartAbandonmentReminder.json
