+++ src/Sharpcode.CartAbandonmentReminder.Web/Scripts/resources/cart-abandonment-reminder-api.js
