+++ src/Sharpcode.CartAbandonmentReminder.Web/package.json
@@ -1,0 +1,17 @@
+{
+    "version": "1.0.0",
+    "name": "asp.net",
+    "private": true,
+    "scripts": {
+        "webpack:dev": "webpack --mode=development",
+        "webpack:watch": "webpack --watch --mode=development",
+        "webpack:build": "webpack --mode=production"
+    },
+    "devDependencies": {
+        "clean-webpack-plugin": "^1.0.0",
+        "css-loader": "^1.0.1",
+        "mini-css-extract-plugin": "^0.4.5",
+        "webpack": "^5.89.0",
+        "webpack-cli": "^4.9.0"
+    }
+}
