--- tests/ReveationLabs.CartAbandonmentReminder.Tests/ReveationLabs.CartAbandonmentReminder.Tests.csproj
@@ -1,29 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-    <TargetFramework>net6.0</TargetFramework>
-    <IsPackable>false</IsPackable>
-  </PropertyGroup>
-
-  <PropertyGroup>
-    <!-- Project is a test project -->
-    <SonarQubeTestProject>true</SonarQubeTestProject>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <PackageReference Include="coverlet.collector" Version="3.1.2">
-      <IncludeAssets>runtime; build; native; contentfiles; analyzers; buildtransitive</IncludeAssets>
-      <PrivateAssets>all</PrivateAssets>
-    </PackageReference>
-    <PackageReference Include="Microsoft.NET.Test.Sdk" Version="16.2.0" />
-    <PackageReference Include="xunit" Version="2.4.0" />
-    <PackageReference Include="xunit.runner.visualstudio" Version="2.4.0" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <ProjectReference Include="..\..\src\ReveationLabs.CartAbandonmentReminder.Core\ReveationLabs.CartAbandonmentReminder.Core.csproj" />
-    <ProjectReference Include="..\..\src\ReveationLabs.CartAbandonmentReminder.Data\ReveationLabs.CartAbandonmentReminder.Data.csproj" />
-    <ProjectReference Include="..\..\src\ReveationLabs.CartAbandonmentReminder.Web\ReveationLabs.CartAbandonmentReminder.Web.csproj" />
-  </ItemGroup>
-
-</Project>
