--- tests/ReveationLabs.CartAbandonmentReminder.Tests/Test.cs
@@ -1,14 +1,0 @@
-using Xunit;
-
-namespace ReveationLabs.CartAbandonmentReminder.Tests
-{
-    [Trait("Category", "Unit")]
-    public class Test
-    {
-        [Fact]
-        public void Run_Test()
-        {
-            Assert.Equal(0, 0);
-        }
-    }
-}
