+++ tests/Sharpcode.CartAbandonmentReminder.Tests/Sharpcode.CartAbandonmentReminder.Tests.csproj
@@ -1,0 +1,29 @@
+<Project Sdk="Microsoft.NET.Sdk">
+
+  <PropertyGroup>
+    <TargetFramework>net8.0</TargetFramework>
+    <IsPackable>false</IsPackable>
+  </PropertyGroup>
+
+  <PropertyGroup>
+    <!-- Project is a test project -->
+    <SonarQubeTestProject>true</SonarQubeTestProject>
+  </PropertyGroup>
+
+  <ItemGroup>
+    <PackageReference Include="coverlet.collector" Version="3.1.2">
+      <IncludeAssets>runtime; build; native; contentfiles; analyzers; buildtransitive</IncludeAssets>
+      <PrivateAssets>all</PrivateAssets>
+    </PackageReference>
+    <PackageReference Include="Microsoft.NET.Test.Sdk" Version="16.2.0" />
+    <PackageReference Include="xunit" Version="2.4.0" />
+    <PackageReference Include="xunit.runner.visualstudio" Version="2.4.0" />
+  </ItemGroup>
+
+  <ItemGroup>
+    <ProjectReference Include="..\..\src\Sharpcode.CartAbandonmentReminder.Core\Sharpcode.CartAbandonmentReminder.Core.csproj" />
+    <ProjectReference Include="..\..\src\Sharpcode.CartAbandonmentReminder.Data\Sharpcode.CartAbandonmentReminder.Data.csproj" />
+    <ProjectReference Include="..\..\src\Sharpcode.CartAbandonmentReminder.Web\Sharpcode.CartAbandonmentReminder.Web.csproj" />
+  </ItemGroup>
+
+</Project>
