+++ tests/Sharpcode.CartAbandonmentReminder.Tests/Test.cs
@@ -1,0 +1,14 @@
+using Xunit;
+
+namespace Sharpcode.CartAbandonmentReminder.Tests
+{
+    [Trait("Category", "Unit")]
+    public class Test
+    {
+        [Fact]
+        public void Run_Test()
+        {
+            Assert.Equal(0, 0);
+        }
+    }
+}
